# textparse: CreatedTimestamp now returns int64 value; optimized proto CT parsing