..instead of *int64. This is as an optimization and ease of use. We already accepted in many places (proto histograms, PRW) that CT (or any timestamp really) 0 means not set. I added todos for normal timestamps too.

Additionally I added trivial fast path for protobuf CreatedTimestamp implementation, was visible on allocs/s in https://github.com/prometheus/prometheus/pull/16046#issuecomment-2663517957

For proto:
```
benchstat v1.txt v3.txt                               
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/textparse
cpu: Apple M1 Pro
                                       │    v1.txt    │               v3.txt               │
                                       │    sec/op    │   sec/op     vs base               │
CreatedTimestampPromProto/case=no-ct-2   22.450n ± 1%   2.232n ± 3%  -90.06% (p=0.002 n=6)
CreatedTimestampPromProto/case=ct-2      14.725n ± 5%   1.874n ± 1%  -87.27% (p=0.002 n=6)
geomean                                   18.18n        2.045n       -88.75%

                                       │   v1.txt   │                 v3.txt                 │
                                       │    B/op    │    B/op     vs base                    │
CreatedTimestampPromProto/case=no-ct-2   16.00 ± 0%    0.00 ± 0%  -100.00% (p=0.002 n=6)
CreatedTimestampPromProto/case=ct-2      8.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
geomean                                  11.31                    ?                      ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                                       │   v1.txt   │                 v3.txt                 │
                                       │ allocs/op  │ allocs/op   vs base                    │
CreatedTimestampPromProto/case=no-ct-2   1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
CreatedTimestampPromProto/case=ct-2      1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
geomean                                  1.000                    ?                      ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```