@bwplotka 2025-02-24
> Why do we need such a long benchmark?

To focus on CT part. Happy to remove.